+++ clover-eclipse/com.atlassian.clover.eclipse.core/src/com/atlassian/clover/eclipse/core/views/nodes/Nodes.java
@@ -298,7 +298,7 @@ public class Nodes {
         TestCaseInfo[] tcis = MetricsScope.TEST_ONLY.getTestCaseInfosFor(method);
         if (tcis != null) {
             for (TestCaseInfo tci : tcis) {
-                if (tci.getId().intValue() == testCaseNode.getTciId()) {
+                if (tci.getId() == testCaseNode.getTciId()) {
                     testCases.add(converter.convert(tci, method));
                     //There can only be one test that maps to a test ID
                     break;
